requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.10
decorator==4.4.0
deprecation==2.0.6
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-JWT-Extended==3.18.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
googlefinance==0.7
idna==2.8
inflection==0.3.1
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.1.0
more-itertools==5.0.0
ndg-httpsclient==0.5.1
numpy==1.16.3
packaging==19.0
pandas==0.24.2
pandas-datareader==0.7.0
parso==0.4.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pmdarima==1.2.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
pyEX==0.1.13
Pygments==2.4.1
PyJWT==1.7.1
pymongo==3.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyramid-arima==0.9.0
python-dateutil==2.8.0
python-engineio==3.6.0
python-socketio==4.0.3
pytz==2019.1
Quandl==3.4.8
requests==2.22.0
scikit-learn==0.21.2
scipy==1.2.2
six==1.12.0
socketIO-client-nexus==0.7.6
sseclient==0.0.24
statsmodels==0.9.0
traitlets==4.3.2
urllib3==1.25.2
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.1
